chore: remove utils
