+++ tests/test_parsers.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from tsbot import utils
+from tsbot import parsers
 
 
 @pytest.mark.parametrize(
@@ -123,7 +123,7 @@ from tsbot import utils
     ),
 )
 def test_parse_data(input_str: str, expected: list[dict[str, str]]) -> None:
-    assert utils.parse_data(input_str) == expected
+    assert parsers.parse_data(input_str) == expected
 
 
 @pytest.mark.parametrize(
@@ -163,7 +163,7 @@ def test_parse_data(input_str: str, expected: list[dict[str, str]]) -> None:
     ),
 )
 def test_parse_line(input_str: str, expected: dict[str, str]) -> None:
-    assert utils.parse_line(input_str) == expected
+    assert parsers.parse_line(input_str) == expected
 
 
 @pytest.mark.parametrize(
@@ -185,28 +185,7 @@ def test_parse_line(input_str: str, expected: dict[str, str]) -> None:
     ),
 )
 def test_parse_value(input_str: str, expected: tuple[str, str]) -> None:
-    assert utils.parse_value(input_str) == expected
-
-
-escape_params = (
-    pytest.param("", "", id="test_empty"),
-    pytest.param("TeamSpeak ]|[ Server", r"TeamSpeak\s]\p[\sServer", id="test_default_server_name"),
-    pytest.param(
-        "Test token with custom set", r"Test\stoken\swith\scustom\sset", id="test_sentance"
-    ),
-    pytest.param("		", r"\t\t", id="test_tabs"),
-    pytest.param("| | | | | | |", r"\p\s\p\s\p\s\p\s\p\s\p\s\p", id="test_space_pipes"),
-)
-
-
-@pytest.mark.parametrize(("input_str", "expected"), escape_params)
-def test_escape(input_str: str, expected: str) -> None:
-    assert utils.escape(input_str) == expected
-
-
-@pytest.mark.parametrize(("expected", "input_str"), escape_params)
-def test_unescape(input_str: str, expected: str) -> None:
-    assert utils.unescape(input_str) == expected
+    assert parsers.parse_value(input_str) == expected
 
 
 @pytest.mark.parametrize(
@@ -282,4 +261,4 @@ def test_unescape(input_str: str, expected: str) -> None:
 def test_parse_args_kwargs(
     input_str: str, expected_args: tuple[str], expected_kwargs: dict[str, str]
 ):
-    assert utils.parse_args_kwargs(input_str) == (expected_args, expected_kwargs)
+    assert parsers.parse_args_kwargs(input_str) == (expected_args, expected_kwargs)
