+++ tsbot/parsers.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from tsbot import encoders
+
 KWARG_INDICATOR = "-"
 
 
@@ -24,7 +26,7 @@ def parse_value(input_str: str) -> tuple[str, str]:
         # Multiple values associated with the key. Making values comma separated
         return key, ",".join(v for v in value.split(f"|{key}="))
 
-    return key, unescape(value) if value else ""
+    return key, encoders.unescape(value) if value else ""
 
 
 def _parse_quoted_arg(unparsed: str) -> tuple[str, str]:
@@ -88,34 +90,3 @@ def parse_args_kwargs(msg: str) -> tuple[tuple[str, ...], dict[str, str]]:
             args.append(value)
 
     return tuple(args), kwargs
-
-
-# https://github.com/benediktschmitt/py-ts3/blob/v2/ts3/escape.py
-
-ESCAPE_MAP = (
-    ("\\", r"\\"),
-    ("/", r"\/"),
-    (" ", r"\s"),
-    ("|", r"\p"),
-    ("\a", r"\a"),
-    ("\b", r"\b"),
-    ("\f", r"\f"),
-    ("\n", r"\n"),
-    ("\r", r"\r"),
-    ("\t", r"\t"),
-    ("\v", r"\v"),
-)
-
-
-def escape(input_str: str) -> str:
-    """Escapes characters that need escaping according to ESCAPE_MAP"""
-    for char, replacement in ESCAPE_MAP:
-        input_str = input_str.replace(char, replacement)
-    return input_str
-
-
-def unescape(input_str: str) -> str:
-    """Undo escaping of characters according to ESCAPE_MAP"""
-    for replacement, char in reversed(ESCAPE_MAP):
-        input_str = input_str.replace(char, replacement)
-    return input_str
